Создает TripController
